kmod: kallsyms_lookup_name kobj_sysfs_ops on older kernels
kobj_sysfs_ops was added to the kernel way back in 2007 with
23b5212cc7422 ("Driver Core: add kobj_attribute handling"), but it was
not exported until much later in 2013 with 29dfe2dc0e8f ("kobject:
export kobj_sysfs_ops"), i.e. kernel v3.14.

Ubuntu Trusty ships with a v3.13 distro kernel that does not export
kobj_sysfs_ops.  In this case, use kallsyms to lookup the symbol if it's
not already defined.

Fixes #709