kpatch: don't complain about missing livepatch .kpatch.checksum
The verify_module_checksum() function reads a kpatch-specific ELF
section to compare on-disk and in-memory kernel modules.  The function
only reports a miscompare if the .kpatch.checksum section actually
exists.  Livepatches don't have such section, so throw away any "Section
'.kpatch.checksum' was not dumped because it does not exist!" warnings
from readelf.

Signed-off-by: Joe Lawrence <joe.lawrence@redhat.com>